+++ .changeset/rich-suns-drive.md
@@ -1,0 +1,5 @@
+---
+'@backstage/backend-openapi-utils': minor
+---
+
+Fixed a Typescript error when trying to use the new OpenAPI server-side generated code.
