+++ packages/backend-openapi-utils/src/index.ts
@@ -30,7 +30,7 @@ export type {
   CookieParameters,
   PathParameters,
 } from './utility';
-export type { ApiRouter } from './router';
+export type { ApiRouter, TypedRouter } from './router';
 export type { PathTemplate } from './types/common';
 export { wrapInOpenApiTestServer, wrapServer } from './testUtils';
 export {
